--- Week-5/MovieApp/app/src/main/java/com/minneydev/movieapp/DetailActivity.kt
@@ -1,15 +1,0 @@
-package com.minneydev.movieapp
-
-import android.content.Intent
-import android.net.Uri
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.minneydev.movieapp.data.Movie
-import com.minneydev.movieapp.ui.DetailAdapter
-import com.squareup.picasso.Picasso
-import kotlinx.android.synthetic.main.activity_detail.*
-
-class DetailActivity : AppCompatActivity() {
-
-}
