+++ Week-5/MovieApp/app/src/main/java/com/minneydev/movieapp/MainActivity.kt
@@ -2,6 +2,9 @@ package com.minneydev.movieapp
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
+import androidx.appcompat.app.AlertDialog
 import androidx.navigation.Navigation
 
 class MainActivity : AppCompatActivity() {
@@ -10,6 +13,26 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         Navigation.findNavController(this, R.id.nav_host_fragment)
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        super.onCreateOptionsMenu(menu)
+        menuInflater.inflate(R.menu.about, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (item.itemId == R.id.aboutMenu) {
+            showInfo()
+        }
+        return true
+    }
+
+    private fun showInfo() {
+        AlertDialog.Builder(this)
+            .setTitle(R.string.alert_title)
+            .setMessage(R.string.alert_message)
+            .create().show()
     }
