+++ Week-5/MovieApp/app/src/main/java/com/minneydev/movieapp/fragments/DetailFragment.kt
@@ -1,4 +1,4 @@
-package com.minneydev.movieapp
+package com.minneydev.movieapp.fragments
 
 import android.content.Intent
 import android.net.Uri
@@ -8,33 +8,24 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.minneydev.movieapp.R
 import com.minneydev.movieapp.data.Movie
 import com.minneydev.movieapp.data.getMoviesArray
 import com.minneydev.movieapp.ui.DetailAdapter
 import com.squareup.picasso.Picasso
-import kotlinx.android.synthetic.main.activity_detail.*
+import kotlinx.android.synthetic.main.fragment_detail.*
 
 class DetailFragment : Fragment() {
 
     companion object {
         lateinit var movie: Movie
 
-        private val ARG_MOVIE = "movie"
-
-        fun newInstance(movie: Movie): DetailFragment {
-            val bundle = Bundle()
-            bundle.putParcelable(ARG_MOVIE, movie)
-            val fragment = DetailFragment()
-            fragment.arguments = bundle
-            return fragment
-        }
     }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_detail, container, false)
     }
 
@@ -42,7 +33,8 @@ class DetailFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         arguments?.let {
-            val args = DetailFragmentArgs.fromBundle(it)
+            val args =
+                DetailFragmentArgs.fromBundle(it)
             movie = getMoviesArray().filter { movie -> movie.title == args.movieString } [0]
         }
 
@@ -61,7 +53,6 @@ class DetailFragment : Fragment() {
 
     private fun displayMovie(movie: Movie?) {
         if (movie != null) {
-//            title = movie.title
             cardTitle.text = movie.title
             Picasso.get()
                 .load(movie.bannerUrl)
