+++ Week-5/MovieApp/app/src/main/java/com/minneydev/movieapp/ui/DetailAdapter.kt
@@ -3,8 +3,7 @@ package com.minneydev.movieapp.ui
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.minneydev.movieapp.DetailActivity
-import com.minneydev.movieapp.DetailFragment
+import com.minneydev.movieapp.fragments.DetailFragment
 import com.minneydev.movieapp.R
 
 class DetailAdapter : RecyclerView.Adapter<DetailViewHolder>() {
@@ -12,8 +11,8 @@ class DetailAdapter : RecyclerView.Adapter<DetailViewHolder>() {
     private val currentMovie = DetailFragment.movie
 
     private val movieInfo = arrayOf(
-        Pair("Summary", currentMovie?.summary),
-        Pair("Release Date", currentMovie?.releaseDate)
+        Pair("Summary", currentMovie.summary),
+        Pair("Release Date", currentMovie.releaseDate)
     )
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DetailViewHolder {
