--- Week-5/MovieApp/app/src/main/res/layout/activity_detail.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/colorGrey"
-    android:orientation="vertical"
-    tools:context=".DetailActivity">
-
-    <ImageView
-        android:id="@+id/moviePoster"
-        android:layout_width="0dp"
-        android:layout_height="@dimen/banner_height"
-        android:layout_marginBottom="@dimen/banner_bottom_margin"
-        android:contentDescription="@string/banner_descrip"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0"
-        app:srcCompat="@drawable/ic_launcher_foreground" />
-
-    <TextView
-        android:id="@+id/cardTitle"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:background="@color/colorGreyDark"
-        android:paddingStart="@dimen/movie_title_padding_start"
-        android:paddingTop="@dimen/movie_title_padding"
-        android:paddingEnd="@dimen/movie_title_padding"
-        android:paddingBottom="@dimen/movie_title_padding"
-        android:text="@string/placeholder"
-        android:textColor="@android:color/white"
-        android:textSize="@dimen/movie_title_size"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/moviePoster" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/searchMovieFab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="@dimen/fab_end_margin"
-        android:layout_marginBottom="@dimen/fab_bottom_margin"
-        android:background="@color/colorPrimary"
-        android:clickable="true"
-        app:backgroundTint="@color/colorPrimary"
-        app:layout_constraintBottom_toBottomOf="@+id/cardTitle"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.954"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.961"
-        app:srcCompat="@android:drawable/ic_menu_search"
-        android:focusable="true" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/detailRecyclerView"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_margin="15dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/cardTitle" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
