+++ Week-5/MovieApp/app/src/main/res/layout/fragment_movie.xml
@@ -9,7 +9,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MovieFragment">
+        tools:context=".fragments.MovieFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/mainRecyclerView"
